+++ src/Waymarked.Api/Program.cs
@@ -1,5 +1,7 @@
 using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.RateLimiting;
 using System.Text;
+using System.Threading.RateLimiting;
 using Waymarked.Api;
 using Waymarked.Api.Data;
 using Waymarked.Api.Email;
@@ -19,9 +21,17 @@ builder.AddNpgsqlDbContext<WaymarkedDbContext>("waymarked");
 builder.Services.AddIdentity<ApplicationUser, IdentityRole>(options =>
     {
         options.User.RequireUniqueEmail = true;
+        options.Lockout.MaxFailedAccessAttempts = 5;
+        options.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(15);
+        options.Lockout.AllowedForNewUsers = true;
     })
     .AddEntityFrameworkStores<WaymarkedDbContext>()
     .AddDefaultTokenProviders();
+
+builder.Services.Configure<DataProtectionTokenProviderOptions>(options =>
+{
+    options.TokenLifespan = TimeSpan.FromHours(24);
+});
 
 builder.Services.ConfigureApplicationCookie(options =>
 {
@@ -43,6 +53,22 @@ builder.Services.ConfigureApplicationCookie(options =>
 });
 
 builder.Services.AddAuthorization();
+
+// Rate limiting (disabled in Test environment to avoid cross-test interference)
+if (!builder.Environment.IsEnvironment("Test"))
+{
+    builder.Services.AddRateLimiter(options =>
+    {
+        options.AddFixedWindowLimiter("forgot-password", limiter =>
+        {
+            limiter.PermitLimit = 3;
+            limiter.Window = TimeSpan.FromMinutes(15);
+            limiter.QueueProcessingOrder = QueueProcessingOrder.OldestFirst;
+            limiter.QueueLimit = 0;
+        });
+        options.RejectionStatusCode = StatusCodes.Status429TooManyRequests;
+    });
+}
 
 // Email
 builder.Services.Configure<SmtpSettings>(builder.Configuration.GetSection("Email"));
@@ -71,6 +97,10 @@ using (var scope = app.Services.CreateScope())
 app.UseExceptionHandler();
 
 app.UseAuthentication();
+if (!app.Environment.IsEnvironment("Test"))
+{
+    app.UseRateLimiter();
+}
 app.UseAuthorization();
 
 if (app.Environment.IsDevelopment())